+++ src/plugin/cursor.js
@@ -83,30 +83,28 @@ export default class CursorPlugin {
     };
 
     /**
-     * @private
      * @param {object} e Mouse move event
      */
     _onMousemove = e => {
         const bbox = this.wavesurfer.container.getBoundingClientRect();
         let y = 0;
         let x = e.clientX - bbox.left;
+        let flip = bbox.right < e.clientX + this.outerWidth(this.displayTime);
 
         if (this.params.showTime && this.params.followCursorY) {
             // follow y-position of the mouse
             y = e.clientY - (bbox.top + bbox.height / 2);
         }
 
-        this.updateCursorPosition(x, y);
+        this.updateCursorPosition(x, y, flip);
     };
 
     /**
-     * @private
      * @returns {void}
      */
     _onMouseenter = () => this.showCursor();
 
     /**
-     * @private
      * @returns {void}
      */
     _onMouseleave = () => this.hideCursor();
@@ -119,9 +117,7 @@ export default class CursorPlugin {
      * @param {object} ws Wavesurfer instance
      */
     constructor(params, ws) {
-        /** @private */
         this.wavesurfer = ws;
-        /** @private */
         this.style = ws.util.style;
         /**
          * The cursor HTML element
@@ -141,8 +137,8 @@ export default class CursorPlugin {
          * @type {?HTMLElement}
          */
         this.displayTime = null;
-        /** @private */
-        this.params = ws.util.extend({}, this.defaultParams, params);
+
+        this.params = Object.assign({}, this.defaultParams, params);
     }
 
     /**
@@ -153,7 +149,7 @@ export default class CursorPlugin {
         this.cursor = this.wrapper.appendChild(
             this.style(
                 document.createElement('cursor'),
-                this.wavesurfer.util.extend(
+                Object.assign(
                     {
                         position: 'absolute',
                         zIndex: this.params.zIndex,
@@ -176,7 +172,7 @@ export default class CursorPlugin {
             this.showTime = this.wrapper.appendChild(
                 this.style(
                     document.createElement('showTitle'),
-                    this.wavesurfer.util.extend(
+                    Object.assign(
                         {
                             position: 'absolute',
                             zIndex: this.params.zIndex,
@@ -196,16 +192,19 @@ export default class CursorPlugin {
             this.displayTime = this.showTime.appendChild(
                 this.style(
                     document.createElement('div'),
-                    this.wavesurfer.util.extend(
+                    Object.assign(
                         {
                             display: 'inline',
                             pointerEvents: 'none',
-                            margin: 'auto'
+                            margin: 'auto',
+                            visibility: 'hidden' // initial value will be hidden just for measuring purpose
                         },
                         this.params.customShowTimeStyle
                     )
                 )
             );
+            // initial value to measure display width
+            this.displayTime.innerHTML = this.formatTime(0);
         }
 
         this.wrapper.addEventListener('mousemove', this._onMousemove);
@@ -237,8 +236,9 @@ export default class CursorPlugin {
      *
      * @param {number} xpos The x offset of the cursor in pixels
      * @param {number} ypos The y offset of the cursor in pixels
+     * @param {boolean} flip Flag to flip duration text from right to left
      */
-    updateCursorPosition(xpos, ypos) {
+    updateCursorPosition(xpos, ypos, flip = false) {
         this.style(this.cursor, {
             left: `${xpos}px`
         });
@@ -255,9 +255,16 @@ export default class CursorPlugin {
             const timeValue =
                 Math.max(0, (xpos / elementWidth) * duration) + scrollTime;
             const formatValue = this.formatTime(timeValue);
+            if (flip) {
+                const textOffset = this.outerWidth(this.displayTime);
+                xpos -= textOffset;
+            }
             this.style(this.showTime, {
                 left: `${xpos}px`,
                 top: `${ypos}px`
+            });
+            this.style(this.displayTime, {
+                visibility: 'visible'
             });
             this.displayTime.innerHTML = `${formatValue}`;
         }
@@ -310,5 +317,21 @@ export default class CursorPlugin {
                 ('000' + Math.floor((time % 1) * 1000)).slice(-3) // milliseconds
             ].join(':')
         );
+    }
+
+    /**
+     * Get outer width of given element.
+     *
+     * @param {DOM} element DOM Element
+     * @returns {number} outer width
+     */
+    outerWidth(element) {
+        if (!element) return 0;
+
+        let width = element.offsetWidth;
+        let style = getComputedStyle(element);
+
+        width += parseInt(style.marginLeft + style.marginRight);
+        return width;
     }
 }