+++ src/plugin/minimap.js
@@ -55,11 +55,12 @@ export default class MinimapPlugin {
     }
 
     constructor(params, ws) {
-        this.params = ws.util.extend(
+        this.params = Object.assign(
             {},
             ws.params,
             {
                 showRegions: false,
+                regionsPluginName: params.regionsPluginName || 'regions',
                 showOverview: false,
                 overviewBorderColor: 'green',
                 overviewBorderSize: 2,
@@ -108,6 +109,7 @@ export default class MinimapPlugin {
         this.renderEvent =
             ws.params.backend === 'MediaElement' ? 'waveform-ready' : 'ready';
         this.overviewRegion = null;
+        this.regionsPlugin = this.wavesurfer[this.params.regionsPluginName];
 
         this.drawer.createWrapper();
         this.createElements();
@@ -127,7 +129,7 @@ export default class MinimapPlugin {
                 ws.container.insertBefore(this.params.container, null);
             }
 
-            if (this.wavesurfer.regions && this.params.showRegions) {
+            if (this.regionsPlugin && this.params.showRegions) {
                 this.regions();
             }
             this.render();